refactor: add seperate file for listeners in projectSocketMount
